# Conflicts:
#	src/aggregator/aggregator/handler/protobuf.go
#	src/query/functions/aggregation/base.go
#	src/query/functions/aggregation/base_test.go
#	src/query/functions/aggregation/count_values.go
#	src/query/functions/aggregation/take.go
#	src/query/models/tags.go
#	src/query/models/tags_test.go
#	src/query/storage/m3/multiresults/multi_fetch_result.go
#	src/query/storage/m3/multiresults/multi_search_result.go
#	src/query/storage/m3/storage.go
#	src/query/storage/m3/types.go
#	src/query/test/comparison.go
#	src/query/ts/series.go
#	src/x/cost/tracker.go

+++ src/dbnode/persist/fs/commitlog/commit_log.go
@@ -632,11 +632,11 @@ func (l *commitLog) writeWait(
 	}
 
 	// Optimistically increment the number of enqueued writes.
-	numEnqueued := atomic.AddInt64(&l.numWritesInQueue, int64(numToEnqueue))
+	numEnqueued := atomic.AddInt64(&l.numWritesInQueue, numToEnqueue)
 
 	// If we exceeded the limit, decrement the number of enqueued writes and bail.
 	if numEnqueued > l.maxQueueSize {
-		atomic.AddInt64(&l.numWritesInQueue, int64(-numToEnqueue))
+		atomic.AddInt64(&l.numWritesInQueue, -numToEnqueue)
 		l.closedState.RUnlock()
 
 		if write.writeBatch != nil {
@@ -677,11 +677,11 @@ func (l *commitLog) writeBehind(
 	}
 
 	// Optimistically increment the number of enqueued writes.
-	numEnqueued := atomic.AddInt64(&l.numWritesInQueue, int64(numToEnqueue))
+	numEnqueued := atomic.AddInt64(&l.numWritesInQueue, numToEnqueue)
 
 	// If we exceeded the limit, decrement the number of enqueued writes and bail.
 	if numEnqueued > l.maxQueueSize {
-		atomic.AddInt64(&l.numWritesInQueue, int64(-numToEnqueue))
+		atomic.AddInt64(&l.numWritesInQueue, -numToEnqueue)
 		l.closedState.RUnlock()
 
 		if write.writeBatch != nil {
